refactor: Wave 6 quality push - god classes, function decomposition, CI consolidation
God class decomposition (3 classes, ~2700 lines extracted into mixins):
- SolarSystemScene (1342→559 lines): extracted SceneEventMixin (406 lines)
  and SceneRenderMixin (303 lines) for event handling and rendering
- ElectrodeVisualization (1116→459 lines): extracted ElectrodeLayersMixin
  (241 lines) for layer/vessel drawing with DRY helper methods
- URDFTextEditor (1173→990 lines): extracted TextEditorDiffMixin (167 lines)
  for diff computation and side-by-side comparison

Function decomposition:
- get_constants_info (123→12 lines): data into module-level _CONSTANTS_METADATA dict
- _render_view_contents decomposed into 6 focused render methods
- _handle_ui_click decomposed into _handle_sidebar_click + _handle_controls_click
- _handle_time_nav_action if/elif chain replaced with dispatch dict

CI workflow consolidation (51→40 workflows):
- Removed 11 orphaned/superseded/duplicate workflows

Test fix:
- Updated test_flow_rate_converter to expect ValueError (DbC precondition)
  instead of ZeroDivisionError for zero molecular weight
