--- .github/workflows/codeql-analysis.yml.disabled
@@ -1,10 +1,0 @@
-name: "CodeQL"
-
-on:
-  workflow_dispatch:
-
-jobs:
-  disabled:
-    runs-on: ubuntu-latest
-    steps:
-      - run: echo "CodeQL is disabled."
