--- .github/workflows/jules-assessment-runner.yml
@@ -1,194 +1,0 @@
-# Jules Assessment Runner Workflow
-#
-# This workflow runs repository assessments on a rotating daily schedule:
-# - Day 1 (Mon/Thu): Assessment A (Architecture & Implementation)
-# - Day 2 (Tue/Fri): Assessment B (Hygiene, Security & Quality)
-# - Day 3 (Wed/Sat): Assessment C (Documentation & Integration)
-# - Day 4 (Sun): Documentation Cleanup Review
-#
-# The assessments are performed by an AI agent reading the prompt templates
-# from docs/assessments/ and generating findings.
-
-name: Jules Assessment Runner
-
-on:
-  # Scheduled runs - adjust time as needed (UTC)
-  schedule:
-    # Run at 2 AM UTC daily
-    - cron: "0 8 * * 0,4"
-
-  # Manual trigger with assessment selection
-  workflow_dispatch:
-    inputs:
-      assessment_type:
-        description: "Which assessment to run"
-        required: true
-        default: "auto"
-        type: choice
-        options:
-          - auto # Uses day-of-week rotation
-          - A # Architecture & Implementation
-          - B # Hygiene, Security & Quality
-          - C # Documentation & Integration
-          - cleanup # Documentation Cleanup
-
-# Prevent multiple assessment runs simultaneously
-concurrency:
-  group: assessment-runner
-  cancel-in-progress: false
-
-jobs:
-  determine-assessment:
-    name: Determine Assessment Type
-    runs-on: ubuntu-latest
-    outputs:
-      assessment: ${{ steps.select.outputs.assessment }}
-      prompt_file: ${{ steps.select.outputs.prompt_file }}
-
-    steps:
-      - name: Determine which assessment to run
-        id: select
-        run: |
-          # If manual trigger with specific type
-          if [[ "${{ github.event.inputs.assessment_type }}" != "" && "${{ github.event.inputs.assessment_type }}" != "auto" ]]; then
-            ASSESSMENT="${{ github.event.inputs.assessment_type }}"
-          else
-            # Auto-select based on day of week (0=Sun, 1=Mon, etc.)
-            DOW=$(date +%u)
-            case $DOW in
-              1|4) ASSESSMENT="A" ;;     # Mon/Thu: Architecture
-              2|5) ASSESSMENT="B" ;;     # Tue/Fri: Hygiene
-              3|6) ASSESSMENT="C" ;;     # Wed/Sat: Documentation
-              7)   ASSESSMENT="cleanup" ;;  # Sun: Cleanup
-            esac
-          fi
-
-          # Map assessment to prompt file
-          case $ASSESSMENT in
-            A) PROMPT_FILE="docs/assessments/Assessment_Prompt_A.md" ;;
-            B) PROMPT_FILE="docs/assessments/Assessment_Prompt_B.md" ;;
-            C) PROMPT_FILE="docs/assessments/Assessment_Prompt_C.md" ;;
-            cleanup) PROMPT_FILE="docs/assessments/Documentation_Cleanup_Prompt.md" ;;
-          esac
-
-          echo "assessment=$ASSESSMENT" >> $GITHUB_OUTPUT
-          echo "prompt_file=$PROMPT_FILE" >> $GITHUB_OUTPUT
-          echo "📋 Selected Assessment: $ASSESSMENT"
-          echo "📄 Prompt File: $PROMPT_FILE"
-
-  run-assessment:
-    name: Run Assessment ${{ needs.determine-assessment.outputs.assessment }}
-    needs: determine-assessment
-    runs-on: ubuntu-latest
-
-    # Skip if triggered by Jules bot (loop prevention)
-    if: github.actor != 'jules-bot'
-
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-        with:
-          fetch-depth: 0 # Full history for analysis
-
-      - name: Read assessment prompt
-        id: read-prompt
-        run: |
-          PROMPT_FILE="${{ needs.determine-assessment.outputs.prompt_file }}"
-          if [[ -f "$PROMPT_FILE" ]]; then
-            echo "✅ Found prompt file: $PROMPT_FILE"
-            # Store first 500 chars as preview (GitHub has output limits)
-            head -c 500 "$PROMPT_FILE" > /tmp/prompt_preview.txt
-          else
-            echo "❌ Prompt file not found: $PROMPT_FILE"
-            exit 1
-          fi
-
-      - name: Generate assessment date stamp
-        id: datestamp
-        run: |
-          DATE=$(date +%Y-%m-%d)
-          echo "date=$DATE" >> $GITHUB_OUTPUT
-          echo "📅 Assessment Date: $DATE"
-
-      # ================================================================
-      # OPTION A: Use Jules AI Agent (if configured)
-      # ================================================================
-      # Uncomment this section if using Jules
-      #
-      # - name: Request Jules Assessment
-      #   uses: jules-ai/assessment-action@v1
-      #   with:
-      #     prompt_file: ${{ needs.determine-assessment.outputs.prompt_file }}
-      #     output_file: docs/assessments/archive/Assessment_${{ needs.determine-assessment.outputs.assessment }}_Results_${{ steps.datestamp.outputs.date }}.md
-      #     model: claude-3  # or gpt-4, etc.
-
-      # ================================================================
-      # OPTION B: Create issue for manual/Copilot review
-      # ================================================================
-      - name: Create assessment tracking issue
-        uses: actions/github-script@v7
-        with:
-          script: |
-            const assessment = '${{ needs.determine-assessment.outputs.assessment }}';
-            const date = '${{ steps.datestamp.outputs.date }}';
-            const promptFile = '${{ needs.determine-assessment.outputs.prompt_file }}';
-
-            const assessmentNames = {
-              'A': 'Architecture & Implementation',
-              'B': 'Hygiene, Security & Quality', 
-              'C': 'Documentation & Integration',
-              'cleanup': 'Documentation Cleanup'
-            };
-
-            const title = `🔍 Assessment ${assessment}: ${assessmentNames[assessment]} - ${date}`;
-
-            const body = `## Scheduled Assessment
-
-            **Type**: Assessment ${assessment} - ${assessmentNames[assessment]}
-            **Date**: ${date}
-            **Prompt File**: \`${promptFile}\`
-
-            ### Instructions for AI Agent
-
-            1. Read the prompt file at \`${promptFile}\`
-            2. Analyze the repository against the assessment criteria
-            3. Generate findings in the specified format
-            4. Post results as a comment on this issue
-            5. Close issue when assessment is complete
-
-            ### Archive Location
-
-            Results should be saved to:
-            \`docs/assessments/archive/Assessment_${assessment}_Results_${date}.md\`
-
-            ---
-
-            <details>
-            <summary>Assessment Prompt Preview</summary>
-
-            See full prompt at: ${promptFile}
-
-            </details>
-
-            /cc @D-sorganization/reviewers
-            `;
-
-            await github.rest.issues.create({
-              owner: context.repo.owner,
-              repo: context.repo.repo,
-              title: title,
-              body: body,
-              labels: ['assessment', `assessment-${assessment.toLowerCase()}`]
-            });
-
-            console.log(`✅ Created assessment issue: ${title}`);
-
-      - name: Summary
-        run: |
-          echo "## Assessment Runner Complete" >> $GITHUB_STEP_SUMMARY
-          echo "" >> $GITHUB_STEP_SUMMARY
-          echo "- **Assessment**: ${{ needs.determine-assessment.outputs.assessment }}" >> $GITHUB_STEP_SUMMARY
-          echo "- **Date**: ${{ steps.datestamp.outputs.date }}" >> $GITHUB_STEP_SUMMARY
-          echo "- **Prompt**: ${{ needs.determine-assessment.outputs.prompt_file }}" >> $GITHUB_STEP_SUMMARY
-          echo "" >> $GITHUB_STEP_SUMMARY
-          echo "An issue has been created for AI agent review." >> $GITHUB_STEP_SUMMARY
